[Rajshri Priya] Add cartapp

+++ book/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render
 from drf_yasg.utils import swagger_auto_schema
+from rest_framework.permissions import BasePermission, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
+# from rest_framework_jwt.authentication import JSONWebTokenAuthentication
+
 from book.models import Book
 from book.serializers import BookSerializer
 from logging_config.logger import get_logger
@@ -12,10 +15,15 @@ logger = get_logger()
 
 # Create your views here.
 class BookAPI(APIView):
+    # authentication_classes = (JSONWebTokenAuthentication,)
+    # permission_classes = [IsAuthenticated]
+
     serializer_class = BookSerializer
 
     @swagger_auto_schema(request_body=BookSerializer, operation_summary='POST Book Created')
     def post(self, request):
+        if not request.user.is_superuser:
+            return Response({"message": "You do not have permission to perform this action."}, status=403)
         try:
             request.data.update({'user': request.user.id})
             serializer = BookSerializer(data=request.data)
@@ -37,6 +45,8 @@ class BookAPI(APIView):
 
     @swagger_auto_schema(request_body=BookSerializer, operation_summary='Book Updated')
     def put(self, request, pk):
+        if not request.user.is_superuser:
+            return Response({"message": "You do not have permission to perform this action."}, status=403)
         try:
             request.data.update({'user': request.user.id})
             book = Book.objects.get(id=pk)
@@ -50,6 +60,8 @@ class BookAPI(APIView):
 
     @swagger_auto_schema(request_body=BookSerializer, operation_summary='Book Deleted')
     def delete(self, request, pk):
+        if not request.user.is_superuser:
+            return Response({"message": "You do not have permission to perform this action."}, status=403)
         try:
             book = Book.objects.get(id=pk, user=request.user)
             book.delete()
@@ -57,3 +69,4 @@ class BookAPI(APIView):
         except Exception as e:
             logger.error(e)
             return Response({"message": str(e)}, status=400)
+
